# PROYECTO/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Hastechno",
        "description": "API para la gestión de productos, pedidos, usuarios, carritos y más para la tienda virtual Hastechno.",
        "contact": {
            "email": "babeltranp@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Servidor de desarrollo local de la API"
        }
    ],
    "paths": {
...
        "/api/consumidores/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener todos los clientes",
                "operationId": "3eee0195c49b8d55057329582187181d",
                "responses": {
                    "200": {
                        "description": "Lista de clientes obtenida correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Clientes obtenidos correctamente."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ClienteConUsuario"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/consumidores/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener un cliente por ID",
                "operationId": "bfcbd5ea4b9cc647e9eb04fd579a1053",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente a obtener.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente encontrado."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ClienteConUsuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clientes/actualizar/cuenta": {
            "patch": {
                "tags": [
                    "Clientes",
                    "Perfil"
                ],
                "summary": "Actualizar parcialmente el perfil del cliente autenticado",
                "operationId": "7fce2387a8e636312ae3b97fd5d687aa",
                "requestBody": {
                    "description": "Datos para actualizar el perfil del cliente. Los campos son opcionales (sometimes).",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Perfil actualizado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado (asociado al usuario autenticado).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clientes/eliminar/cuenta": {
            "delete": {
                "tags": [
                    "Clientes",
                    "Perfil"
                ],
                "summary": "Eliminar la cuenta del cliente autenticado",
                "operationId": "6ec17e3175b7b0aa0d53a962570bd7c1",
                "responses": {
                    "200": {
                        "description": "Cuenta eliminada correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cuenta eliminada correctamente."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado (asociado al usuario autenticado).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/consumidores/clientes/actualizar/{id}": {
            "patch": {
                "tags": [
                    "Clientes",
                    "Administración"
                ],
                "summary": "Actualizar un cliente por ID (solo para administradores)",
                "operationId": "8a33bda645157c4d32da4a86254b11f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos para actualizar el cliente. Los campos son opcionales (sometimes).",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente actualizado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/consumidores/clientes/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Clientes",
                    "Administración"
                ],
                "summary": "Eliminar un cliente por ID (solo para administradores)",
                "operationId": "9efd04eacd75d4a50320c981e8337d34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente eliminado exitosamente."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClienteConUsuario": {
                "title": "Cliente con Usuario",
                "description": "Representa un cliente con la información de su usuario asociada.",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Cliente"
                    },
                    {
                        "properties": {
                            "usuario": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT con el prefijo 'Bearer ' (ej: 'Bearer eyJ0eXAiOiJKV1Q...') para acceder a los endpoints protegidos.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}